Add profile stats to the CPU profiler
